requirements.txt
# AI REASONING
langchain
langgraph
langchain-openai
langchain-community
# langfuse

# DJANGO PURPOSE
django>=5.0,<5.1
djangorestframework    #building APIs
django-cors-headers    #To allow frontend to talk to backend
django-environ         #for reading .env file 


# databse and storage
psycopg[binary]
redis

# --- Asynchronous Server (For Pipecat/WebSockets) ---
uvicorn[standard]           # High-performance ASGI server
daphne                      # Django's official ASGI interface (optional, but good)


requests
python-dotenv
protobuf==4.25.3         
celery
pipecat-ai
livekit
silero